perubahan database, removed bad code
database post -> content diubah ke type long text biar bisa nmpung isi dari file, backend di landing udah di pindah ke controller

+++ src/database/migrations/2020_11_15_074245_create_results_table.php
@@ -17,7 +17,7 @@ class CreateResultsTable extends Migration
             $table->id();
             $table->foreignId('search_id');
             $table->foreignId('post_id');
-            $table->integer('similarity');
+            $table->float('similarity');
             $table->timestamps();
         });
     }
